Test the list hosts endpoint
