+++ server/service/service_hosts_test.go
@@ -3,7 +3,6 @@ package service
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/WatchBeam/clock"
 	"github.com/fleetdm/fleet/v4/server/contexts/viewer"
@@ -15,34 +14,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestListHosts(t *testing.T) {
-	ds := mysql.CreateMySQLDS(t)
-	defer ds.Close()
-
-	svc := newTestService(ds, nil, nil)
-
-	hosts, err := svc.ListHosts(test.UserContext(test.UserAdmin), fleet.HostListOptions{})
-	assert.Nil(t, err)
-	assert.Len(t, hosts, 0)
-
-	storedTime := time.Now().UTC()
-
-	_, err = ds.NewHost(context.Background(), &fleet.Host{
-		Hostname:        "foo",
-		SeenTime:        storedTime,
-		DetailUpdatedAt: time.Now(),
-		LabelUpdatedAt:  time.Now(),
-		PolicyUpdatedAt: time.Now(),
-	})
-	require.NoError(t, err)
-
-	hosts, err = svc.ListHosts(test.UserContext(test.UserAdmin), fleet.HostListOptions{})
-	require.NoError(t, err)
-	require.Len(t, hosts, 1)
-	format := "%Y-%m-%d %HH:%MM:%SS %Z"
-	assert.Equal(t, storedTime.Format(format), hosts[0].SeenTime.Format(format))
-}
 
 func TestDeleteHost(t *testing.T) {
 	ds := mysql.CreateMySQLDS(t)
